move Status to Utilities
