+++ RunControlGenerator/src/main/java/org/genevaers/runcontrolgenerator/repositorybuilders/XMLBuilder.java
@@ -28,9 +28,9 @@ import java.util.Collection;
 import org.genevaers.repository.Repository;
 import org.genevaers.repository.data.InputReport;
 import org.genevaers.runcontrolgenerator.configuration.RunControlConfigration;
-import org.genevaers.runcontrolgenerator.utility.Status;
 import org.genevaers.utilities.GersFile;
 import org.genevaers.utilities.GersFilesUtils;
+import org.genevaers.utilities.Status;
 
 import com.google.common.flogger.FluentLogger;
 import com.google.common.flogger.StackSize;
